scx: separate out the bpf and userspace sdt_task code
This separates out the sdt_task code into a definitions header and the
functionality header that is needed for BPF.  Since we only use the
definitions for userspace to read the stats we can just include that in
common.h with the appropriate arena helper.
